style(gauges): unify panel header, padding, and combo styling
Both gauges now share a title header, inner padding, and the display
ComboBox styling via helpers in gauges/mod.rs, so sibling gauges render
consistently. The geo-position gauge gains the same title as the
orientation gauge.
